Response refactoring in Service implementations.
All methods now return a Response object instead of a JSON. This way we package anything that might be coming as a response from the server inside a certain context. Current responses are BadRequest, ForbiddenRequest, AuthFailed. This way we frame the response our server gives into a correct context so the frontend can easily recongize whether there has been an issue or not.